requirements.txt
kafka-python
redis
streamlit
pandas
pydeck
geographiclib
